Rename to response_mapper
