+++ state_app/test/api/response/mapper/response_mapper_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:state_app/api/response/mapper/Mapper.dart';
+import 'package:state_app/api/response/mapper/response_mapper.dart';
 import 'package:state_app/model/artwork.dart';
 import 'package:state_app/model/common_name.dart';
 import 'package:state_app/model/recent_track.dart';
